tasks: Add clientbuffer module installation
This commit does two things:

1. Adds tasks file to install clientbuffer third-party module
2. Splits up task files across multiple smaller files

The clientbuffer playbook installs ZNC development libraries, clones the
git repo with the third-party module, compiles it, installs it in place,
and restarts ZNC. I tested this manually on a CentOS 7.7 server and it
worked without an issue.

Splitting up the files into smaller pieces is to help with readability,
and also allows people to opt out of the clientbuffer module if they
don't want it, but do want the ZNC install pieces.

Signed-off-by: Justin W. Flory <git@jwf.io>