Add cron script for LetsEncrypt/Certbot renewals
This commit adds a cron script to execute a `certbot renew` command. The
cron script is then installed to `/etc/cron.weekly/` so it is executed
once a week.

This is intended to address the certbot post-renewal script not running.
Somehow the cron job was lost. This puts the cron job into config
management so this problem doesn't happen.

cc: @akarshanbiswas

Signed-off-by: Justin W. Flory <git@jwf.io>